Removig compiler warning 2 commits ago caused segfault.
The segfault was issued because toSet was a single operation,
but the new approach using list.begin and list.end can not be used with veinsharedcomponent directly.
list.begin and list.end are placed in different lists, because value returns a casted copy of the actual variable.
To work around this problem its necessary to store the value in a tmp variable on stack first.
Signed-off-by: bhamacher <b.hamacher@zera.de>